refactor(commands): extract command descriptors into shared module
- Create command-descriptors.ts with shared CommandDescriptor and SpotifyCommandDescriptor types
- Simplify registry.tsx to re-export types and expose a flat COMMAND_RENDERERS map
- Simplify spotify-registry.tsx with the same flat renderer map pattern
- Update TerminalInput to import descriptors from the new shared module
- Remove redundant COMMAND_DEFS/SPOTIFY_COMMAND_DEFS array-to-map transformations
